+++ src/acmap_point.cpp
@@ -3,6 +3,7 @@
 #include "acmap_titers.h"
 #include "acmap_plotspec.h"
 #include "acmap_point.h"
+#include "acmap_sequences.h"
 
 // Regular details
 std::string AcPoint::get_type() const { return type; }
@@ -15,6 +16,7 @@ std::string AcPoint::get_id() const { return id; }
 std::string AcPoint::get_species() const { return species; }
 int AcPoint::get_group() const { return group; }
 std::string AcPoint::get_sequence() const { return sequence; }
+std::vector<SeqInsertion> AcPoint::get_sequence_insertions() const { return sequence_insertions; }
 std::vector<std::string> AcPoint::get_clade() const { return clade; }
 std::vector<std::string> AcPoint::get_annotations() const { return annotations; }
 std::vector<std::string> AcPoint::get_labids() const { return labids; }
@@ -34,6 +36,7 @@ void AcPoint::set_id( std::string value ){ id = value; }
 void AcPoint::set_species( std::string value ){ species = value; }
 void AcPoint::set_group( int value ){ group = value; }
 void AcPoint::set_sequence( std::string value ){ sequence = value; }
+void AcPoint::set_sequence_insertions( std::vector<SeqInsertion> value ){ sequence_insertions = value; }
 void AcPoint::set_clade( std::vector<std::string> value ){ clade = value; }
 void AcPoint::set_annotations( std::vector<std::string> value ){ annotations = value; }
 void AcPoint::set_labids( std::vector<std::string> value ){ labids = value; }
@@ -67,6 +70,8 @@ bool AcPoint::isdefault(std::string attribute) {
     return(group == 0);
   } else if (attribute == "sequence") {
     return(sequence == "");
+  } else if (attribute == "sequence_insertions") {
+    return(sequence_insertions.size() == 0);
   } else if (attribute == "id") {
     return(id == "");
   } else if (attribute == "extra") {